% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polygon.R
\name{buffer_polygon}
\alias{buffer_polygon}
\title{Enlarge/Buffer a Polygon}
\usage{
buffer_polygon(polygon, buffer_distance)
}
\arguments{
\item{polygon}{An sf object representing the polygon to be buffered.}

\item{buffer_distance}{Numeric. The buffer distance in decimal degrees (arc degrees).}
}
\value{
An sf object representing the buffered polygon.
}
\description{
This function enlarges a polygon by applying a buffer.
}
