% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_coordinates}
\alias{clean_coordinates}
\title{Clean Coordinates of Presence/Absence Data}
\usage{
clean_coordinates(
  data,
  sf_poly,
  overlapping = TRUE,
  round_digits = FALSE,
  n_digits = 0
)
}
\arguments{
\item{data}{A data frame containing presence/absence data with columns "decimalLongitude" and "decimalLatitude".}

\item{sf_poly}{A spatial polygon representing the boundaries within which coordinates should be kept.}

\item{overlapping}{Logical indicating whether points overlapping the polygon should be kept (TRUE) or removed (FALSE).}

\item{round_digits}{Logical indicating whether to round coordinates.}

\item{n_digits}{Number of digits to round the coordinates to, if 'round_digits' is TRUE.}
}
\value{
A cleaned data frame containing presence/absence data with valid coordinates.
}
\description{
This function cleans coordinates of presence/absence data by removing NA coordinates, rounding coordinates if specified, removing duplicated points, and removing points outside specified spatial polygon boundaries.
}
\details{
This function takes a data frame containing presence/absence data with longitude and latitude coordinates, a spatial polygon representing boundaries within which to keep points, and parameters for rounding coordinates and handling duplicated points. It returns a cleaned data frame with valid coordinates within the specified boundaries.
}
\keyword{internal}
