% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_coords_layer}
\alias{create_coords_layer}
\title{Create Geographic Coordinate Layers}
\usage{
create_coords_layer(layers, study_area = NULL, scale_layers = FALSE)
}
\arguments{
\item{layers}{Raster or stack of raster layers to derive geographic extent and resolution.}

\item{study_area}{Spatial object for masking output layers.}

\item{scale_layers}{Logical indicating if scaling is applied. Default is FALSE.}
}
\value{
Raster stack with layers lon and lat.
}
\description{
Generates raster layers for longitude and latitude from given raster data,
applies optional scaling, and restricts the output to a specified spatial mask.
}
