% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{downloadActionButton}
\alias{downloadActionButton}
\title{Create a Download Action Button}
\usage{
downloadActionButton(
  outputId,
  label = "Download",
  icon = NULL,
  width = NULL,
  status = NULL,
  outline = FALSE,
  ...
)
}
\arguments{
\item{outputId}{The output ID for the button.}

\item{label}{The label text displayed on the button. Default is "Download".}

\item{icon}{The icon to be displayed on the button. Default is NULL.}

\item{width}{The width of the button. Default is NULL.}

\item{status}{The status of the button. Default is NULL.}

\item{outline}{Logical indicating whether to use outline style for the button. Default is FALSE.}

\item{...}{Additional parameters to be passed to the actionButton function.}
}
\value{
Returns a download action button with the specified parameters.
}
\description{
This function generates a download action button that triggers the download of a file when clicked.
}
\keyword{internal}
