% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_plot.R
\name{export_plot_server}
\alias{export_plot_server}
\title{Server Logic for Export Plot Functionality}
\usage{
export_plot_server(id, exported_plot)
}
\value{
No return value, this function is used for its side effects within a Shiny app.
}
\description{
Sets up server-side functionality for exporting plots, including creating a modal dialog
for user input on export preferences (height, width, format) and processing the download.
}
\keyword{internal}
