% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export_plot.R
\name{export_plot_ui}
\alias{export_plot_ui}
\title{Create UI for Export Plot Button}
\usage{
export_plot_ui(id)
}
\value{
Returns an actionButton for use in a Shiny UI that triggers plot export modal when clicked.
}
\description{
This function generates a UI element (action button) for exporting plots.
The button is styled to be minimalistic, featuring only a download icon.
}
\keyword{internal}
