% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_input_area.R
\name{file_input_area_server}
\alias{file_input_area_server}
\title{Server-side Logic for Custom File Input}
\usage{
file_input_area_server(id)
}
\value{
A reactive expression that returns a data frame containing information about the uploaded files, or NULL if no files have been uploaded.
}
\description{
Processes the file input from the UI component created by \code{\link{file_input_area_ui}} and provides access to the uploaded file data.
}
\keyword{internal}
