% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_file_input_area.R
\name{file_input_area_ui}
\alias{file_input_area_ui}
\title{Custom File Input UI}
\usage{
file_input_area_ui(
  id,
  label = "Input text: ",
  multiple = FALSE,
  accept = NULL,
  width = NULL,
  button_label = "Browse...",
  icon_name = NULL
)
}
\value{
A Shiny UI object that can be added to a Shiny application.
}
\description{
Creates a customized file input area in a Shiny application. The file input is designed to be visually distinct and supports features such as multiple file selection and file type restrictions.
}
\keyword{internal}
