% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_cv_plot}
\alias{generate_cv_plot}
\title{Generate Cross-Validation Plot}
\usage{
generate_cv_plot(data)
}
\arguments{
\item{data}{Dataframe containing cross-validation results.}
}
\value{
Returns a ggplot object representing the cross-validation plot.
}
\description{
This function generates a cross-validation plot based on evaluation metrics.
}
\keyword{internal}
