% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_prediction_plot}
\alias{generate_prediction_plot}
\title{Generate Prediction Plot}
\usage{
generate_prediction_plot(
  prediction_layer,
  pa_points,
  legend_label,
  non_study_area_mask,
  coords
)
}
\arguments{
\item{prediction_layer}{Raster prediction layer.}

\item{pa_points}{Presence/absence points.}

\item{legend_label}{Label for the legend.}

\item{non_study_area_mask}{Spatial polygon representing the non study areas.}
}
\value{
Returns a ggplot object representing the world prediction plot.
}
\description{
This function generates a plot based on prediction raster layers and presence/absence points.
}
\keyword{internal}
