% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_absences.R
\name{generate_pseudo_absences}
\alias{generate_pseudo_absences}
\title{Generate pseudo-absence points based on presence points, covariates, and study area polygon}
\usage{
generate_pseudo_absences(
  presences,
  study_area,
  raster_stack,
  predictor_variables,
  coords = c("decimalLongitude", "decimalLatitude"),
  sp_thin_dist = NULL,
  attempts = 100
)
}
\arguments{
\item{presences}{Data frame containing presence points.}

\item{study_area}{Spatial polygon defining the study area ('sf').}

\item{raster_stack}{SpatRaster containing covariate data.}

\item{predictor_variables}{Name of the predictor variables selected for this species.}

\item{coords}{Character vector specifying the column names for latitude and longitude.}

\item{sp_thin_dist}{Distance in kilometers that you want the occurrences to be separated by.}

\item{attempts}{Number of attempts to generate exact pseudo-absences.}
}
\value{
Data frame containing both presence and pseudo-absence points.
}
\description{
This function generates pseudo-absence points within the study area.
}
