% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_world_prediction_plot}
\alias{generate_world_prediction_plot}
\title{Generate World Prediction Plot}
\usage{
generate_world_prediction_plot(
  prediction_layer,
  pa_points,
  legend_label,
  global_land_mask
)
}
\arguments{
\item{prediction_layer}{Raster prediction layer.}

\item{pa_points}{Presence/absence points.}

\item{legend_label}{Label for the legend.}

\item{global_land_mask}{Spatial polygon representing global land mask.}
}
\value{
Returns a ggplot object representing the world prediction plot.
}
\description{
This function generates a world prediction plot based on prediction raster layers and presence/absence points.
}
\keyword{internal}
