% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_covariate_names}
\alias{get_covariate_names}
\title{Get Covariate Names}
\usage{
get_covariate_names(file_path)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}
}
\value{
A character vector containing the names of covariates.
}
\description{
This function extracts the names of covariates from a ZIP file containing covariate layers.
}
\details{
This function extracts the names of covariates from a ZIP file containing covariate layers.
}
\keyword{internal}
