% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFprTpr}
\alias{getFprTpr}
\title{Compute specificity and sensitivity}
\usage{
getFprTpr(actuals, predictedScores, threshold = 0.5)
}
\description{
Compute specificity and sensitivity
}
\details{
This function was obtained from the InformationValue R package (https://github.com/selva86/InformationValue).
}
\keyword{internal}
