% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layers.R
\name{global_mask}
\alias{global_mask}
\title{Apply Global Mask to Raster Layers}
\usage{
global_mask(layers, sf_poly)
}
\arguments{
\item{layers}{A SpatRaster object or stack of raster layers to be processed.}

\item{sf_poly}{A spatial polygon (`sf` or `sp` object) used to mask the raster layers.}
}
\value{
A `Raster*` object representing the masked raster layers.
}
\description{
This function crops and extends raster layers to a global extent defined by longitude
from -180 to 180 and latitude from -90 to 90, then applies a mask based on a provided
spatial polygon to remove areas outside the polygon.
}
