% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glossa_export}
\alias{glossa_export}
\title{Export Glossa Model Results}
\usage{
glossa_export(
  sp = NULL,
  mods = NULL,
  time = NULL,
  fields = NULL,
  model_data = FALSE,
  fr = FALSE,
  prob_cut = FALSE,
  varimp = FALSE,
  layer_format = "tif",
  prediction_results = NULL,
  presence_absence_list = NULL,
  other_results = NULL,
  pa_cutoff = NULL
)
}
\description{
This function exports various types of Glossa model results, including native range predictions, suitable habitat predictions, model data, variable importance, functional response results, and presence/absence probability cutoffs. It generates raster files for prediction results, CSV files for model data and variable importance, and CSV files for functional response results. Additionally, it creates a CSV file for presence/absence probability cutoffs if provided.
}
\keyword{internal}
