% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layers.R
\name{layer_mask}
\alias{layer_mask}
\title{Apply Polygon Mask to Raster Layers}
\usage{
layer_mask(layers, sf_poly)
}
\arguments{
\item{layers}{A SpatRaster object or stack of raster layers to be processed.}

\item{sf_poly}{A spatial polygon (`sf` or `sp` object) used to mask the raster layers.}
}
\value{
A `Raster*` object representing the masked raster layers.
}
\description{
This function crops and extends raster layers to a sf_poly extent (bbox) defined by longitude
and latitude then applies a mask based on a provided spatial polygon to remove areas outside the polygon.
}
