% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{misClassError}
\alias{misClassError}
\title{Misclassification Error}
\usage{
misClassError(actuals, predictedScores, threshold = 0.5)
}
\description{
Misclassification Error
}
\details{
This function was obtained from the InformationValue R package (https://github.com/selva86/InformationValue).
}
\keyword{internal}
