% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{optimalCutoff}
\alias{optimalCutoff}
\title{Compute the optimal probability cutoff score}
\usage{
optimalCutoff(
  actuals,
  predictedScores,
  optimiseFor = "misclasserror",
  returnDiagnostics = FALSE
)
}
\description{
Compute the optimal probability cutoff score
}
\details{
This function was obtained from the InformationValue R package (\url{https://github.com/selva86/InformationValue}).
}
\keyword{internal}
