% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cv_folds_points}
\alias{plot_cv_folds_points}
\title{Plot cross-validation fold assignments}
\usage{
plot_cv_folds_points(data, polygon = NULL)
}
\arguments{
\item{data}{Dataframe with columns: 'decimalLongitude', 'decimalLatitude', 'pa' and 'fold'.}

\item{polygon}{An sf object representing the inverted study area.}
}
\value{
A ggplot object showing point color-coded by cv fold and shaped by presence/absence.
}
\description{
Plot cross-validation fold assignments
}
