% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cv_metrics}
\alias{plot_cv_metrics}
\title{Plot cross-validation metrics}
\usage{
plot_cv_metrics(data)
}
\arguments{
\item{data}{Dataframe containing cross-validation results.}
}
\value{
Returns a ggplot object representing the cross-validation plot.
}
\description{
This function generates a cross-validation radial plot based on evaluation metrics.
}
\keyword{internal}
