% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{predict_bart}
\alias{predict_bart}
\title{Make Predictions Using a BART Model}
\usage{
predict_bart(bart_model, layers, cutoff = NULL)
}
\arguments{
\item{bart_model}{A BART model object obtained from fitting BART (bart).}

\item{layers}{A SpatRaster object containing environmental covariates for prediction.}

\item{cutoff}{An optional cutoff value for determining potential presences.}
}
\value{
A SpatRaster containing the mean, median, standard deviation, and quantiles
of the posterior predictive distribution.
}
\description{
This function makes predictions using a Bayesian Additive Regression Trees (BART) model
on a stack of environmental covariates.
}
