% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_covariate_layers}
\alias{process_covariate_layers}
\title{Process Covariate Layers}
\usage{
process_covariate_layers(layers, ocean_mask)
}
\arguments{
\item{layers}{A list of raster layers representing covariates.}

\item{ocean_mask}{A spatial polygon representing the ocean boundaries.}
}
\value{
A list containing processed raster layers for each covariate.
}
\description{
This function processes covariate layers by cropping them to the globe extent, extending them to cover the entire globe, and removing land values.
}
\details{
This function takes a list of raster layers representing covariates and a spatial polygon representing ocean boundaries. It crops each layer to the globe extent, extends it to cover the entire globe, and removes land values using the ocean mask.
}
\keyword{internal}
