% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{read_fit_layers_zip}
\alias{read_fit_layers_zip}
\title{Read and validate fit layers zip}
\usage{
read_fit_layers_zip(file_path, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the zip file containing layers to fit the model.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
A raster stack of layers if the layers pass validation criteria, NULL otherwise.
}
\description{
This function validates a zip file containing layers. It checks if the layers have the same number of files, CRS (Coordinate Reference System), and resolution.
}
\keyword{internal}
