% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_glossa_layers}
\alias{read_glossa_layers}
\title{Function to load covariate layers from ZIP files and perform validation}
\usage{
read_glossa_layers(file_path)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}
}
\value{
A list containing raster layers for each covariate.
}
\description{
Function to load covariate layers from ZIP files and perform validation
}
\details{
This function loads covariate layers from a ZIP file, verifies their spatial characteristics, and returns them as a list of raster layers.
}
