% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_files.R
\name{read_glossa_projections_layers}
\alias{read_glossa_projections_layers}
\title{Load Covariate Layers from ZIP Files}
\usage{
read_glossa_projections_layers(file_path)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}
}
\value{
A list containing raster layers for each covariate.
}
\description{
This function loads covariate layers from a ZIP file, verifies their spatial characteristics, and returns them as a list of raster layers.
}
\details{
This function expects the ZIP file to contain subdirectories for each covariate, with raster files stored within these directories. All covariates should have the same number of files and the same spatial characteristics (e.g., CRS, resolution).
}
