% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{read_last_layer}
\alias{read_last_layer}
\title{Read last layer for each environmental variable}
\usage{
read_last_layer(file_path, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the zip file containing layers to fit the model.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}

\item{all_layers}{Logical. If TRUE all layers will be loaded, if FALSE only the last layer of each environmental variable.}
}
\value{
A raster stack of layers if the layers pass validation criteria, NULL otherwise.
}
\description{
Read last layer for each environmental variable for previsualization purposes
}
\keyword{internal}
