% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{read_layers_zip}
\alias{read_layers_zip}
\title{Load Covariate Layers from ZIP Files}
\usage{
read_layers_zip(file_path, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
A list containing raster layers for each covariate.
}
\description{
This function loads covariate layers from a ZIP file, verifies their spatial characteristics, and returns them as a list of raster layers.
}
\keyword{internal}
