% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{read_projections_layers}
\alias{read_projections_layers}
\title{Load Covariate Layers from ZIP Files}
\usage{
read_projections_layers(file_path)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}
}
\value{
A list containing raster layers for each covariate.
}
\description{
This function loads covariate layers from a ZIP file, verifies their spatial characteristics, and returns them as a list of raster layers.
}
\keyword{internal}
