% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_coordinates.R
\name{remove_close_points}
\alias{remove_close_points}
\title{Remove Close Points}
\usage{
remove_close_points(
  data,
  sp_thin_dist,
  coords = c("decimalLongitude", "decimalLatitude")
)
}
\arguments{
\item{data}{A data frame containing the spatial data to be thinned. This data frame must include columns for longitude and latitude.}

\item{sp_thin_dist}{A numeric value specifying the minimum distance (in kilometers) that must separate the data points.}

\item{coords}{A character vector of length 2, specifying the column names in `data` that contain the longitude and latitude values, respectively. Defaults to `c("decimalLongitude", "decimalLatitude")`.}
}
\value{
A data frame that contains the subset of the original data, filtered to ensure that no two points are closer together than the specified `sp_thin_dist`.
}
\description{
Filters out data points that are closer together than a specified distance.
}
