% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_coordinates.R
\name{remove_duplicate_points}
\alias{remove_duplicate_points}
\title{Remove duplicated points from a dataframe}
\usage{
remove_duplicate_points(x, coords = c("decimalLongitude", "decimalLatitude"))
}
\arguments{
\item{x}{A dataframe object.}

\item{coords}{Names of the coordinate columns used for identifying duplicate points.}
}
\value{
A dataframe without duplicated points.
}
\description{
This function removes duplicated points from a dataframe object based on specified coordinate column names.
}
