% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_coordinates.R
\name{remove_points_poly}
\alias{remove_points_poly}
\title{Remove points from a spatial dataframe that fall within or outside a polygon.}
\usage{
remove_points_poly(
  x,
  study_area,
  overlapping = TRUE,
  coords = c("decimalLongitude", "decimalLatitude")
)
}
\arguments{
\item{x}{A dataframe object with rows representing points.}

\item{study_area}{Polygon object (sf object) to define the region for point removal.}

\item{overlapping}{Logical indicating whether to remove points within (TRUE) or outside (FALSE) the polygon.}

\item{coords}{Character vector specifying the column names for longitude and latitude.}
}
\value{
A data.frame containing the filtered points.
}
\description{
This function removes points from a spatial dataframe that either fall within or outside the specified polygon,
depending on the 'overlapping' parameter.
}
