% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_presence_absence.R
\name{remove_points_polygon}
\alias{remove_points_polygon}
\title{Remove Points Inside or Outside a Polygon}
\usage{
remove_points_polygon(
  df,
  polygon,
  overlapping = FALSE,
  coords = c("decimalLongitude", "decimalLatitude")
)
}
\arguments{
\item{df}{A dataframe object with rows representing points.}

\item{polygon}{An sf polygon object defining the region for point removal.}

\item{overlapping}{Logical indicating whether points overlapping the polygon should be removed (TRUE) or kept (FALSE).}

\item{coords}{Character vector specifying the column names for longitude and latitude. Default is c("decimalLongitude", "decimalLatitude").}
}
\value{
A dataframe containing the filtered points.
}
\description{
This function removes points from a dataframe based on their location relative to a specified polygon.
}
