% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{response_curve_bart}
\alias{response_curve_bart}
\title{Calculate Response Curve Using BART Model}
\usage{
response_curve_bart(bart_model, data, predictor_names)
}
\arguments{
\item{bart_model}{A BART model object obtained from fitting BART ('dbarts::bart').}

\item{data}{A data frame containing the predictor variables (the design matrix) used in the BART model.}

\item{predictor_names}{A character vector containing the names of the predictor variables.}
}
\value{
A list containing a data frame for each independent variable with mean, 2.5th percentile, 97.5th percentile, and corresponding values of the variables.
}
\description{
This function calculates the response curve (functional responses) using a Bayesian Additive Regression Trees (BART) model.
}
