% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glossa_analysis.R
\name{run_glossa_analysis}
\alias{run_glossa_analysis}
\title{Main Analysis Function for Glossa Package}
\usage{
run_glossa_analysis(
  pa_files = NULL,
  historical_files = NULL,
  future_files = NULL,
  future_scenario_names = NULL,
  round_digits = FALSE,
  n_round_digits = 1,
  scale_layers = FALSE,
  native_range = NULL,
  suitable_habitat = NULL,
  other_analysis = NULL,
  seed = NA,
  waiter = NULL
)
}
\arguments{
\item{pa_files}{A list of file paths containing presence-absence data.}

\item{historical_files}{A list of file paths containing historical environmental layers.}

\item{future_files}{A list of file paths containing future scenario environmental layers.}

\item{future_scenario_names}{A vector of names corresponding to future scenarios.}

\item{round_digits}{Logical; if TRUE, coordinates will be rounded to a specified number of digits.}

\item{n_round_digits}{Integer; number of digits to round coordinates to if `round_digits` is TRUE.}

\item{scale_layers}{Logical; if TRUE, covariate layers will be scaled based on historical data.}

\item{native_range}{A vector of scenarios ('historical', 'past', 'future') where native range modeling should be performed.}

\item{suitable_habitat}{A vector of scenarios ('historical', 'past', 'future') where habitat suitability modeling should be performed.}

\item{other_analysis}{A vector of additional analyses to perform (e.g., 'variable_importance', 'response_curve', 'cross_validation').}

\item{seed}{Optional; an integer seed for reproducibility of results.}

\item{waiter}{Optional; a waiter instance to update progress in a Shiny application.}
}
\value{
A list containing structured outputs from each major section of the analysis including model data, predictions,
variable importance scores, and habitat suitability assessments.
}
\description{
This function wraps all the analysis that the Glossa package performs. It processes presence-absence data,
environmental covariates, and performs species distribution modeling and prediction under historical, past, and future scenarios.
}
