% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glossa_app.R
\name{run_glossa_app}
\alias{run_glossa_app}
\title{Run Glossa Shiny App}
\usage{
run_glossa_app(maxRequestSize = NULL)
}
\arguments{
\item{maxRequestSize}{To increase the maximum file size allowed for uploading in Shiny (default 5Mb)}
}
\description{
This function launches the Glossa Shiny web application.
}
\details{
The Glossa Shiny app provides an interactive interface for users to access Glossa functionalities.
}
\examples{
\dontrun{
run_glossa_app()
}
}
