% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparkvalueBox}
\alias{sparkvalueBox}
\title{Create a Sparkline Value Box}
\usage{
sparkvalueBox(
  title,
  sparkline_data,
  description,
  type = "line",
  box_color = "white",
  width = 4,
  elevation = 0,
  ...
)
}
\arguments{
\item{title}{The title or heading of the value box.}

\item{sparkline_data}{The data used to generate the sparkline plot.}

\item{description}{A short description or additional information displayed below the value box.}

\item{type}{The type of sparkline plot to generate. Default is "line".}

\item{box_color}{The background color of the value box.}

\item{width}{The width of the value box. Default is 4.}

\item{elevation}{The elevation of the value box. Default is 0.}

\item{...}{Additional parameters to be passed to the sparkline function.}
}
\value{
Returns a custom value box with the specified parameters.
}
\description{
This function creates a custom value box with a sparkline plot embedded in it.
}
\keyword{internal}
