% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{specificity}
\alias{specificity}
\title{Calculate the specificity for a given logit model}
\usage{
specificity(actuals, predictedScores, threshold = 0.5)
}
\description{
Calculate the specificity for a given logit model
}
\details{
This function was obtained from the InformationValue R package (\url{https://github.com/selva86/InformationValue}).
}
\keyword{internal}
