% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_files.R
\name{validate_extent_poly}
\alias{validate_extent_poly}
\title{Validate Extent Polygon}
\usage{
validate_extent_poly(file_input, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the polygon file containing the extent.}
}
\value{
A spatial object representing the extent if the file is valid, NULL otherwise.
}
\description{
This function validates a polygon file containing the extent. It checks if the file has the correct format.
}
\details{
This function validates the format of a polygon file containing the extent. It checks if the file has the correct format.
}
\keyword{internal}
