% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_fit_layers_zip}
\alias{validate_fit_layers_zip}
\title{Validate Layers Zip}
\usage{
validate_fit_layers_zip(file_input, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the zip file containing layers.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates a zip file containing layers. It checks if the layers have the same number of files, CRS (Coordinate Reference System), and resolution.
}
\details{
This function expects that each subfolder within the zip file represents a covariate, and each covariate contains one or more raster files. It checks if the layers within each covariate have the same CRS and resolution.
}
\keyword{internal}
