% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_fit_proj_layers}
\alias{validate_fit_proj_layers}
\title{Validate Fit and Projection Layers}
\usage{
validate_fit_proj_layers(fit_layers_path, proj_path, show_modal = FALSE)
}
\arguments{
\item{fit_layers_path}{Path to the ZIP file containing fit layers.}

\item{proj_path}{Path to the ZIP file containing projection layers.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates fit and projection layers by checking their CRS, resolution, and covariates.
}
\details{
This function compares fit and projection layers to ensure they have the same CRS, resolution, and covariates.
}
\keyword{internal}
