% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{validate_fit_projection_layers}
\alias{validate_fit_projection_layers}
\title{Validate Fit and Projection Layers}
\usage{
validate_fit_projection_layers(
  fit_layers_path,
  proj_layers_path,
  show_modal = FALSE
)
}
\arguments{
\item{fit_layers_path}{Path to the ZIP file containing fit layers.}

\item{proj_layers_path}{Path to the ZIP file containing projection layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates fit and projection layers by checking their covariates.
}
\keyword{internal}
