% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_hist_fut_layers}
\alias{validate_hist_fut_layers}
\title{Validate Historical and Future Layers}
\usage{
validate_hist_fut_layers(hist_path, fut_path)
}
\arguments{
\item{hist_path}{Path to the ZIP file containing historical layers.}

\item{fut_path}{Path to the ZIP file containing future layers.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates historical and future layers by checking their CRS, resolution, and covariates.
}
\details{
This function compares historical and future layers to ensure they have the same CRS, resolution, and covariates.
}
\keyword{internal}
