% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{validate_layers_zip}
\alias{validate_layers_zip}
\title{Validate Layers Zip}
\usage{
validate_layers_zip(file_path, timestamp_mapping = NULL, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the ZIP file containing environmental layers.}

\item{timestamp_mapping}{Optional. A vector with the timestamp mapping of the environmental layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates a ZIP file containing environmental layers. It checks if the layers have the same number of files, CRS (Coordinate Reference System), and resolution.
}
\keyword{internal}
