% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{validate_pa_fit_time}
\alias{validate_pa_fit_time}
\title{Validate match between P/A files and fit layers}
\usage{
validate_pa_fit_time(pa_data, fit_layers_path, show_modal = FALSE)
}
\arguments{
\item{pa_data}{P/A uploaded dataframe.}

\item{fit_layers_path}{Path to the ZIP file containing fit layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
TRUE if the layers and csv files pass validation criteria, FALSE otherwise.
}
\description{
This function validates the match between the time periods of the P/A files and the environmental layers
}
\keyword{internal}
