% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_presences_absences_csv}
\alias{validate_presences_absences_csv}
\title{Validate presences/absences CSV file}
\usage{
validate_presences_absences_csv(file_input, show_modal = FALSE)
}
\arguments{
\item{file_path}{The file path to the CSV file.}
}
\value{
TRUE if the file has the expected columns and formats, FALSE otherwise.
}
\description{
This function validates a CSV file containing presences and absences data for species occurrences. It checks if the file has the expected columns and formats.
}
\details{
This function validates the format of a CSV file containing presence/absence data. It checks if the file has the expected columns and formats. If the "pa" column is missing, it assumes the presence/absence column and adds it with default values.
}
\keyword{internal}
