% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{validate_projection_layers_zip}
\alias{validate_projection_layers_zip}
\title{Validate Projection Layers Zip}
\usage{
validate_projection_layers_zip(file_path, show_modal = FALSE)
}
\arguments{
\item{file_path}{Path to the zip file containing projection layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
TRUE if the layers pass validation criteria, FALSE otherwise.
}
\description{
This function validates a zip file containing projection layers. It checks if the layers have the same number of files, CRS (Coordinate Reference System), and resolution.
}
\keyword{internal}
