% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{variable_importance}
\alias{variable_importance}
\title{Variable Importance in BART model}
\usage{
variable_importance(bart_model)
}
\arguments{
\item{bart_model}{A BART model object.}
}
\value{
A numeric vector containing the name of the variable and its corresponding importance value. The variables are sorted such that the first variable listed is the one that has the greatest influence.
}
\description{
Calculate variable importance, measured as the proportion of total branches used for a given variable.
}
