% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{youdensIndex}
\alias{youdensIndex}
\title{Calculate Youden's index}
\usage{
youdensIndex(actuals, predictedScores, threshold = 0.5)
}
\description{
Calculate Youden's index
}
\details{
This function was obtained from the InformationValue R package (\url{https://github.com/selva86/InformationValue}).
}
\keyword{internal}
