% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{fit_bart_model}
\alias{fit_bart_model}
\title{Fit a BART Model Using Environmental Covariate Layers}
\usage{
fit_bart_model(y, x, seed = NULL)
}
\arguments{
\item{y}{Vector indicating presence (1) or absence (0).}

\item{x}{Dataframe with same number of rows as the length of the vector `y` with the covariate values.}

\item{seed}{Random seed.}
}
\value{
A BART model object.
}
\description{
This function fits a Bayesian Additive Regression Trees (BART) model using
presence/absence data and environmental covariate layers.
}
