% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_polygon.R
\name{invert_polygon}
\alias{invert_polygon}
\title{Invert a Polygon}
\usage{
invert_polygon(polygon, bbox = NULL)
}
\arguments{
\item{polygon}{An sf object representing the polygon to be inverted.}

\item{bbox}{Optional. An sf or bbox object representing the bounding box. If NULL, the bounding box of the input polygon is used.}
}
\value{
An sf object representing the inverted polygon.
}
\description{
This function inverts a polygon by calculating the difference between the bounding box and the polygon.
}
