% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_polygon.R
\name{invert_polygon}
\alias{invert_polygon}
\title{Invert a polygon with respect to the world.
This function inverts a polygon with respect to the entire world.}
\usage{
invert_polygon(sf_poly)
}
\arguments{
\item{sf_poly}{An sf polygon to be inverted.}
}
\value{
An sf polygon representing the inverted area.
}
\description{
Invert a polygon with respect to the world.
This function inverts a polygon with respect to the entire world.
}
