% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_polygon.R
\name{invert_polygon}
\alias{invert_polygon}
\title{Invert a polygon}
\usage{
invert_polygon(sf_poly, bbox = NULL)
}
\arguments{
\item{sf_poly}{An sf polygon to be inverted.}

\item{bbox}{Optionally provide the resulting bounding box.}
}
\value{
An sf polygon representing the inverted area.
}
\description{
This function inverts a polygon.
}
