% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_layers.R
\name{layer_mask}
\alias{layer_mask}
\title{Apply Polygon Mask to Raster Layers}
\usage{
layer_mask(layers, study_area)
}
\arguments{
\item{layers}{A stack of raster layers (`SpatRaster` object) to be processed.}

\item{study_area}{A spatial polygon (`sf` object) used to mask the raster layers.}
}
\value{
A `SpatRaster` object representing the masked raster layers.
}
\description{
This function crops and extends raster layers to a study area extent (bbox) defined by longitude
and latitude then applies a mask based on a provided spatial polygon to remove areas outside the polygon.
}
