% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{pa_optimal_cutoff}
\alias{pa_optimal_cutoff}
\title{Optimal Cutoff for Presence-Absence Prediction}
\usage{
pa_optimal_cutoff(y, x, model, seed = NULL)
}
\arguments{
\item{y}{Vector indicating presence (1) or absence (0).}

\item{x}{Dataframe with same number of rows as the length of the vector `y` with the covariate values.}

\item{model}{A BART model object.}

\item{seed}{Random seed for reproducibility.}
}
\value{
The optimal cutoff value for presence-absence prediction.
}
\description{
This function calculates the optimal cutoff for presence-absence prediction using a BART model.
}
