% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bart_functions.R
\name{pa_optimal_cutoff}
\alias{pa_optimal_cutoff}
\title{Optimal Cutoff for Presence-Absence Prediction}
\usage{
pa_optimal_cutoff(pa_coords, layers, model, seed = NULL)
}
\arguments{
\item{pa_coords}{Data frame with coordinates of presences and absences, and a column indicating presence or absence.}

\item{layers}{A list of layer names to extract from the raster stack.}

\item{model}{A BART model object.}

\item{seed}{Random seed for reproducibility.}
}
\value{
The optimal cutoff value for presence-absence prediction.
}
\description{
This function calculates the optimal cutoff for presence-absence prediction using a BART model.
}
