% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_glossa_pa}
\alias{read_glossa_pa}
\title{Function to load presence/absence data from CSV files and perform validation}
\usage{
read_glossa_pa(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file containing presence/absence data.}
}
\value{
A data frame containing the loaded presence/absence data.
}
\description{
Function to load presence/absence data from CSV files and perform validation
}
\details{
This function loads presence/absence data from a CSV file, performs validation to ensure required columns are present, and adds a default value for the presence/absence column if not present.
}
