% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{read_layers_zip}
\alias{read_layers_zip}
\title{Load Covariate Layers from ZIP Files}
\usage{
read_layers_zip(
  file_path,
  extend = TRUE,
  first_layer = FALSE,
  show_modal = FALSE
)
}
\arguments{
\item{file_path}{Path to the ZIP file containing covariate layers.}

\item{extend}{If TRUE it will take the largest extent, if FALSE the smallest.}

\item{first_layer}{If TRUE it will return only the layers from the first timestamp.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
A list containing raster layers for each covariate.
}
\description{
This function loads covariate layers from a ZIP file, verifies their spatial characteristics, and returns them as a list of raster layers.
}
\keyword{internal}
