% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_presence_absence.R
\name{remove_duplicate_points}
\alias{remove_duplicate_points}
\title{Remove Duplicated Points from a Dataframe}
\usage{
remove_duplicate_points(df, coords = c("decimalLongitude", "decimalLatitude"))
}
\arguments{
\item{df}{A dataframe object with each row representing one point.}

\item{coords}{A character vector specifying the names of the coordinate columns used for identifying duplicate points. Default is c("decimalLongitude", "decimalLatitude").}
}
\value{
A dataframe without duplicated points.
}
\description{
This function removes duplicated points from a dataframe based on specified coordinate columns.
}
