% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_files.R
\name{validate_pa_fit_time}
\alias{validate_pa_fit_time}
\title{Validate Match Between Presence/Absence Files and Fit Layers}
\usage{
validate_pa_fit_time(pa_data, fit_layers_path, show_modal = FALSE)
}
\arguments{
\item{pa_data}{Data frame containing the presence/absence data with a 'timestamp' column.}

\item{fit_layers_path}{Path to the ZIP file containing fit layers.}

\item{show_modal}{Optional. Logical. Whether to show a modal notification for warnings. Default is FALSE.}
}
\value{
TRUE if the timestamps match the fit layers, FALSE otherwise.
}
\description{
This function validates whether the time periods of the presence/absence data match the environmental layers.
}
\keyword{internal}
